Integrate IPC pluggable infrastructure
This commit integrates the IPC pluggable infrastructure
and an IPC (ZeroMQ) module that can be loaded using such
a mechanism.

It extends the vsm script to accept the '--ipc-module'
option to switch between the default stdin/stodut module
and the ZeroMQ IPC module.

It includes tests for the IPC module to check that the
plugin can be correctly loaded and that signals are
properly sent and processed by VSM.

Signed-off-by: Luis Araujo <luis.araujo@collabora.co.uk>