Support subconditions and rule monitors
Subconditions (conditions which have at least one level of conditions
above them) automatically cause the creation of rule monitors when all
ancestor conditions are met.

While a monitor exists, the subcondition must be met by `start`
milliseconds and maintained until `stop` milliseconds after the creation
of the monitor. The `start` and `stop` times are specified in the rule
file by keywords of the same names.

If any ancestor condition becomes false, the monitor is cancelled
immediately.

This also adds support for 'parallel' blocks as a type of wrapper block.
Subconditions must be wrapped in a wrapper block.

Based on code originally written by Shane Fagan