Add catapult logging format
This commit adds support to write the log file in the
Trace Event Format (the format used by the Catapult
Trace Viewer).

It records only the signal emissions (incoming/outgoing)
as Instant Events and converting the time to microseconds
since that is the trace viewer timestamps format.

This log format can be selected passing 'catapult' to the
--log-format command line option.

Signed-off-by: Luis Araujo <luis.araujo@collabora.co.uk>