Fix floating point values
This commit fixes the floating point values (they were not recognized),
and it improves checking the values types raising a ValueError if the
value is invalid.

It also makes sure that a string value is only between "" or '' (and not
a combination of both) and that empty lines are not processed.

Signed-off-by: Luis Araujo <luis.araujo@collabora.co.uk>