Add IPC class to use instead of global module functions
Rather than using global module functions such as open, close, send
and receive, define an IPC class and create an instance of it to
communicate with the IPC.  This improves the design by creating a
better defined interface to the IPC modules, and also enables adding
support for multiple IPC modules in follow-up changes.

Consolidate the default functions into a StdioIPC class to use stdin
and stdout, and instantiate that by default with the --ipc-module
option is not used.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>