vsm: do not abort when signal is None
When an IPC module receives a signal it doesn't know how to decode, it
should ignore it and its receive() method should return None.  When
this happens, skip the unknown signal in the VSM main loop.  If it
should have been decoded and passed to the VSM but failed, the
relevant IPC module can decide to raise an exception to actually
abort.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>