tests: move ZeroMQ tests to a separate class
Split the TestVSM class into TestVSMDebug which deals with DebugIPC
and TestVSMZeroMQ which deals with ZeromqIPC.

Also split the TestVSM class to separate the run_vsm logic from the
test cases, which are now in VSMTestCases.  Create new sub-classes of
VSMTestCases which defines which IPC module to use.

This removes a lot of conditionals and makes the code easier to
maintain and expand with new IPC modules that need to run a different
set of test cases.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>