vsm: convert to vsm.py so it can be imported
Rename `vsm' to `vsm.py' so it can be imported.  Also remove the
shebang line and change file permissions to non-executable.

Further improvement would be to create a `vsm' directory to turn this
into a Python package rather than a single module, with vsmlib and ipc
inside that directory.  This patch is the minimal changes required to
import vsm and add some customisation.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>