ipc.IPCList: directly pass list of module objects
Rather than passing a list of class names to IPCList, directly pass a
list of module objects to make it more generic.  In particular, this
allows the caller to create any list of module objects with arbitrary
constructor arguments.

Adjust init_ipc() accordingly in vsm.py to create a list of modules
using the class names provide on the command line.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>