vsm: cancel delayed signals when exiting main loop
Delayed signals are run by sleeping in a separate thread.  When the
main VSM loop exits, they need to be aborted otherwise the process
will be waiting for each of them to time out.

When scheduling a delayed signal, create an Event object and wait on
it with a timeout.  Add it to a global set of events, so that when the
main loop exits it can iterate and set all the pending events.  Make
the waiting threads detect wether the event was set or timed out, to
only emit delayed signals when they actually time out.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>